// apps/desktop/src/types/generated/NetworkInterface.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NetworkInterfaceKind } from "./NetworkInterfaceKind";

/**
 * Interfaz de red visible en el sistema.
 */
export type NetworkInterface = { name: string, kind: NetworkInterfaceKind, mac_address: string | null, ipv4_addresses: Array<string>, is_up: boolean, };